Moved the old classes back to package me.leejay.jenkins.dateparameter
