+++ src/main/java/me/leejay/jenkins/dateparameter/DateParameterDefinition.java
@@ -4,8 +4,6 @@ import hudson.Extension;
 import hudson.model.ParameterDefinition;
 import hudson.model.ParameterValue;
 import hudson.util.FormValidation;
-import me.leejay.jenkins.DateParameterValue;
-import me.leejay.jenkins.StringLocalDateValue;
 import net.sf.json.JSONObject;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
