+++ src/main/java/me/leejay/jenkins/dateparameter/DateParameterValue.java
@@ -1,4 +1,4 @@
-package me.leejay.jenkins;
+package me.leejay.jenkins.dateparameter;
 
 import hudson.AbortException;
 import hudson.EnvVars;
